+++ addon-beta/CHANGELOG.md
@@ -1,4 +1,4 @@
-##### *GENERATED FILE [2026-08-08 20:08 UTC] — do not edit by hand.*
+##### *GENERATED FILE [2026-08-08 20:41 UTC] — do not edit by hand.*
 
 # Changelog